server/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements/requirements.in
#
amqp==5.1.1
    # via kombu
asgiref==3.7.1
    # via django
async-timeout==4.0.2
    # via redis
beautifulsoup4==4.12.2
    # via -r requirements/requirements.in
billiard==3.6.4.0
    # via celery
celery==5.2.7
    # via -r requirements/requirements.in
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
dj-database-url==2.0.0
    # via -r requirements/requirements.in
django==4.2.1
    # via
    #   -r requirements/requirements.in
    #   dj-database-url
    #   djangorestframework
djangorestframework==3.14.0
    # via -r requirements/requirements.in
idna==3.4
    # via requests
kombu==5.2.4
    # via celery
lxml==4.9.2
    # via -r requirements/requirements.in
prompt-toolkit==3.0.38
    # via click-repl
psycopg2-binary==2.9.6
    # via -r requirements/requirements.in
python-decouple==3.8
    # via -r requirements/requirements.in
python-dotenv==1.0.0
    # via -r requirements/requirements.in
pytz==2023.3
    # via
    #   celery
    #   djangorestframework
redis==4.5.4
    # via -r requirements/requirements.in
requests==2.29.0
    # via -r requirements/requirements.in
six==1.16.0
    # via click-repl
soupsieve==2.4.1
    # via beautifulsoup4
sqlparse==0.4.4
    # via django
typing-extensions==4.6.2
    # via
    #   asgiref
    #   dj-database-url
urllib3==1.26.16
    # via requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit
